+++ feature/feature-base/src/main/java/org/gampiot/robokide/feature/base/ui/RobokActivity.kt
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 
 import com.google.android.material.appbar.MaterialToolbar
 
-import org.gampiot.robokide.feature.base.R
+import org.gampiot.robokide.feature.res.R
 import org.gampiot.robokide.feature.manage.file.requestPermission
 import org.gampiot.robokide.feature.util.getBackPressedClickListener
